requirements.txt
alpaca-trade-api==3.2.0
alpha-vantage==2.3.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
websocket-client==1.6.4